requirements.txt
gradio>=4.0.0
trafilatura>=1.6.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
anthropic>=0.25.0

# Telegram bot (bot/)
python-telegram-bot>=21.0
notion-client>=2.2
